Fixup target ID type

+++ X6000Plus/kern_hwlibs.hpp
@@ -15,7 +15,7 @@ class HWLibs {
     bool processKext(KernelPatcher &patcher, size_t id, mach_vm_address_t slide, size_t size);
 
     private:
-    int targetDeviceId;
+    uint64_t targetDeviceId;
 
     static char *wrapGetMatchProperty(void);
 };
